Patch 6.1.070 (depends on 6.1.060) Problem:    Compiler warning for signed/unsigned mismatch. (Mike Williams) Solution:   Add a typecast to int. Files:	    src/ops.c

+++ src/ops.c
@@ -1665,7 +1665,7 @@ op_delete(oap)
 
 	    /* Break a tab only when it's included in the area. */
 	    if (gchar_pos(&oap->end) == '\t'
-					  && oap->end.coladd < oap->inclusive)
+				     && (int)oap->end.coladd < oap->inclusive)
 	    {
 		/* save last line for undo */
 		if (u_save((linenr_T)(oap->end.lnum - 1),
